revise google analytics, add fb, twitter sharing

+++ NaviUtil/Place.m
@@ -225,7 +225,7 @@
     if (nil == name || name.length < 1)
         return true;
     
-    int location, length, matchedCount, i;
+    long location, length, matchedCount, i;
     float matchedRate = 0;
     location = [self.name rangeOfString:name].location;
     length = [self.name rangeOfString:name].length;
